requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Pillow==10.1.0
bcrypt==4.1.1
python-dotenv==1.0.0
WTForms==3.1.1
email-validator==2.1.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
